pyproject.toml
[tool.poetry]
name = "automatedciliameasurements"
version = "0.1.0"
description = "Pipeline to measure cilia, cells, and which cilia are close to which cells from microscopy images using CellProfiler and Python"
authors = ["sneha <sneha.shinde350+0@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
numpy = "^1.21.2"
pandas = "^1.3.3"
sklearn = "^0.0"
Pillow-PIL = "^0.1dev"
bokeh = "^2.4.2"
matplotlib = "^3.5.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-docstrings = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
